Moved AsyncStringDownloader to separate directory
