+++ Minecraft Version Downloader/Core/Downloaders/VersionInfoFileListDownloaders/FilteredVersionListDownloader.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Minecraft_Server_Downloader.Core.Downloaders.StringDownloaders;
 using Minecraft_Server_Downloader.Structures;
 
 namespace Minecraft_Server_Downloader.Core.Downloaders.VersionInfoFileListDownloaders
