+++ Minecraft Version Downloader/Core/Downloaders/VersionInfoFileListDownloaders/IncrementalVersionListDownloader.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Minecraft_Server_Downloader.Core.Downloaders.StringDownloaders;
 using Minecraft_Server_Downloader.Structures;
 
 namespace Minecraft_Server_Downloader.Core.Downloaders.VersionInfoFileListDownloaders
